// api/types/UpdateCourseRequest.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { UpdateCourseDayRequest } from "./UpdateCourseDayRequest.ts";

export type UpdateCourseRequest = {
    /**
     * @minLength 1
     * @type string
    */
    title: string;
    /**
     * @type string | undefined
    */
    description?: string;
    /**
     * @type string | undefined
    */
    imageUrl?: string;
    /**
     * @type integer, int64
    */
    price: number;
    /**
     * @type integer, int32
    */
    durationDays: number;
    /**
     * @type integer | undefined, int64
    */
    categoryId?: number;
    /**
     * @type array
    */
    courseDays: UpdateCourseDayRequest[];
};