// api/types/UpdateCourseDayRequest.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { UpdateDayExerciseRequest } from "./UpdateDayExerciseRequest.ts";

export type UpdateCourseDayRequest = {
    /**
     * @type integer, int32
    */
    dayNumber: number;
    /**
     * @type string | undefined
    */
    dayDescription?: string;
    /**
     * @type array
    */
    dayExercises: UpdateDayExerciseRequest[];
};